requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.2
djangorestframework==3.15.2
coreapi==2.3.3
idna==3.10
psycopg2==2.9.10
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
django-cors-headers==4.6.0
pillow
opencv-python
numpy
keras
python-multipart
tensorflow
skin-tone-classifier
gdown
requests
dlib
